SS-120: changes to the store fields and code cleanup.
