--- web/src/types/Assignment/jsonTypes/index.ts
@@ -1,10 +1,0 @@
-import {} from '../../common';
-
-export interface placeHolderType {
-    "Id": string,
-    "fullNm": string,
-    "phoneNumberTxt": string,
-    "additionalProp1": {},
-    "additionalProp2": {},
-    "additionalProp3": {}
-}
