+++ web/src/types/ManageTypes/index.ts
@@ -1,13 +1,12 @@
 import {} from '../common';
 import {} from '../DutyRoster/jsonTypes';
 
-export interface placeHolderInfoType {
-
-    "ID": string,
-    "Role": string,
-    "First Name": string,
-    "Last Name": string,
-    "Name": string,
-    "Index": number
+export interface assignmentTypeInfoType {
+    code: string;
+    concurrencyToken?: number;
+    id: number;
+    locationId: number;
+    type: string;
+    sortOrder: number;
 }
